pyproject.toml
[project]
name = "trails"
version = "0.1.0"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.135.3",
    "uvicorn[standard]>=0.42.0",
    "sqlalchemy>=2.0",
    "alembic>=1.14",
    "asyncpg>=0.30",
    "pgvector>=0.4.2",
    "celery>=5.6",
    "redis>=7.4.0",
    "httpx>=0.28.1",
    "pydantic-settings>=2.6",
    "python-jose[cryptography]>=3.3",
    "cryptography>=46.0.6",
    "exponent-server-sdk>=2.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.setuptools]
py-modules = []
